+++ packages/client/src/utils/helpers.ts
@@ -60,3 +60,22 @@ export const sortGames = (
       );
   }
 };
+
+export const getFilterOptions = (checkedFormInputs: HTMLInputElement[]) =>
+  checkedFormInputs.reduce((prev, curr) => {
+    switch (curr.type) {
+      case "radio":
+        return {
+          ...prev,
+          [curr.name]: +curr.value,
+        };
+      case "checkbox":
+        return {
+          ...prev,
+          [curr.name]: prev[curr.name]
+            ? [...prev[curr.name], +curr.value]
+            : [+curr.value],
+        };
+    }
+    return {};
+  }, {} as any);
